Update to have json editing capability
