+++ src/frontend/shortlist/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@popperjs/core": "^2.11.6",
+    "axios": "^1.1.3",
     "bootstrap": "^5.2.2",
     "vue": "^3.2.38",
     "vue-router": "^4.1.5"
