+++ src/frontend/shortlist/src/App.vue
