+++ src/frontend/shortlist/src/views/StageView.vue
@@ -1,3 +1,13 @@
+<script>
+
+import SimCardDemo from '../components/school/simple/SimCardDemo.vue';
+
+    export default {
+    components: {
+        SimCardDemo,
+    },
+}
+</script>
 <template>
-    <div>Stage</div>
+    <SimCardDemo></SimCardDemo>
 </template>
